requirements.txt
paho-mqtt==1.6.1
pandas==2.0.3
openpyxl==3.1.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3 